move client into own file + fix compile issues
